refactor: 패키지명 annotation 으로 변경 (#99)
